+++ src/clojure_lsp/parser.clj
@@ -57,10 +57,10 @@
   (let [forms (find-forms zloc (fn [loc]
                                  (in-range?
                                   (-> loc z/node meta) pos)))
-        disconsider-reader-macro? (and (some #(= '? (z/sexpr %)) forms)
-                                    (> (count forms) 1))]
+        disconsider-reader-macro? (and (some #(= "?" (z/string %)) forms)
+                                       (> (count forms) 1))]
     (if disconsider-reader-macro?
-      (last (filter (complement (comp #(= '? %) z/sexpr)) forms))
+      (last (filter (complement (comp #(= "?" %) z/string)) forms))
       (last forms))))
 
 (defn find-top-forms-in-range
@@ -90,9 +90,9 @@
         (throw e)))))
 
 (defn loc-at-pos [code row col]
-  (-> code
-      safe-zloc-of-string
-      (find-last-by-pos {:row row :col col :end-row row :end-col col})))
+  (some-> code
+          safe-zloc-of-string
+          (find-last-by-pos {:row row :col col :end-row row :end-col col})))
 
 (defn safe-loc-at-pos [text row col]
   (try